Move LNLSZoom fucntions to EPICSNState
This can be done because all NState objetcs also use these funcions
and because limits can be defined in configuration files
